pyproject.toml
[tool.poetry]
name = "rl-bot"
version = "0.1.0"
description = ""
authors = ["Kyle Peront <kaperont@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
discord-py = "^2.3.2"
django-environ = "^0.11.2"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
